+++ cupy/fft/fft.py
@@ -3,7 +3,141 @@ import six
 import numpy as np
 
 import cupy
-from cupy import cufft
+from cupy.cuda import cufft
+
+
+def _convert_dtype(a, value_type):
+    if value_type != 'R2C':
+        if a.dtype in [np.float16, np.float32]:
+            return a.astype(np.complex64)
+        elif a.dtype not in [np.complex64, np.complex128]:
+            return a.astype(np.complex128)
+    else:
+        if a.dtype in [np.complex64, np.complex128]:
+            return a.real
+        elif a.dtype == np.float16:
+            return a.astype(np.float32)
+        elif a.dtype not in [np.float32, np.float64]:
+            return a.astype(np.float64)
+    return a
+
+
+def _cook_shape(a, s, axes, value_type):
+    if s is None:
+        return a
+    if not hasattr(s, '__iter__'):
+        s = (s,)
+    if (value_type == 'C2R') and (s[-1] is not None):
+        s = list(s)
+        s[-1] = s[-1] // 2 + 1
+    for sz, axis in zip(s, axes):
+        if (sz is not None) and (sz != a.shape[axis]):
+            shape = list(a.shape)
+            if shape[axis] > sz:
+                index = [slice(None)] * a.ndim
+                index[axis] = slice(0, sz)
+                a = a[index]
+            else:
+                index = [slice(None)] * a.ndim
+                index[axis] = slice(0, shape[axis])
+                shape[axis] = sz
+                z = cupy.zeros(shape, a.dtype.char)
+                z[index] = a
+                a = z
+    return a
+
+
+def _convert_fft_type(a, value_type):
+    if value_type == 'C2C' and a.dtype == np.complex64:
+        return cufft.CUFFT_C2C
+    elif value_type == 'R2C' and a.dtype == np.float32:
+        return cufft.CUFFT_R2C
+    elif value_type == 'C2R' and a.dtype == np.complex64:
+        return cufft.CUFFT_C2R
+    elif value_type == 'C2C' and a.dtype == np.complex128:
+        return cufft.CUFFT_Z2Z
+    elif value_type == 'R2C' and a.dtype == np.float64:
+        return cufft.CUFFT_D2Z
+    else:
+        return cufft.CUFFT_Z2D
+
+
+def _exec_fft(a, direction, value_type, norm, axis, out_size=None):
+    fft_type = _convert_fft_type(a, value_type)
+
+    if axis % a.ndim != a.ndim - 1:
+        a = a.swapaxes(axis, -1)
+
+    if a.base is not None:
+        a = a.copy()
+
+    plan = cufft.Plan1d(a.shape[-1] if out_size is None else out_size,
+                        fft_type, a.size // a.shape[-1])
+    out = plan.get_output_array(a)
+    plan.fft(a, out, direction)
+
+    sz = out.shape[-1]
+    if fft_type == cufft.CUFFT_R2C or fft_type == cufft.CUFFT_D2Z:
+        sz = a.shape[-1]
+    if norm is None:
+        if direction == cufft.CUFFT_INVERSE:
+            out /= sz
+    else:
+        out /= cupy.sqrt(sz)
+
+    if axis % a.ndim != a.ndim - 1:
+        out = out.swapaxes(axis, -1)
+
+    return out
+
+
+def _fft_c2c(a, direction, norm, axes):
+    for axis in axes:
+        a = _exec_fft(a, direction, 'C2C', norm, axis)
+    return a
+
+
+def _fft(a, s, axes, norm, direction, value_type='C2C'):
+    if norm not in (None, 'ortho'):
+        raise ValueError('Invalid norm value %s, should be None or \"ortho\".'
+                         % norm)
+
+    if s is not None:
+        for n in s:
+            if (n is not None) and (n < 1):
+                raise ValueError(
+                    "Invalid number of FFT data points (%d) specified." % n)
+
+    if (s is not None) and (axes is not None) and len(s) != len(axes):
+        raise ValueError("Shape and axes have different lengths.")
+
+    a = _convert_dtype(a, value_type)
+    if axes is None:
+        if s is None:
+            dim = a.ndim
+        else:
+            dim = len(s)
+        axes = [i for i in six.moves.range(-dim, 0)]
+    elif isinstance(axes, np.compat.integer_types):
+        axes = (axes,)
+    a = _cook_shape(a, s, axes, value_type)
+
+    if value_type == 'C2C':
+        a = _fft_c2c(a, direction, norm, axes)
+    elif value_type == 'R2C':
+        a = _exec_fft(a, direction, value_type, norm, axes[-1])
+        a = _fft_c2c(a, direction, norm, axes[:-1])
+    else:
+        a = _fft_c2c(a, direction, norm, axes[:-1])
+        if isinstance(s, np.compat.integer_types):
+            out_size = s
+        elif (s is None) or (s[-1] is None):
+            out_size = a.shape[axes[-1]] * 2 - 2
+        else:
+            out_size = s[-1]
+        a = _exec_fft(a, direction, value_type, norm, axes[-1], out_size)
+
+    return a
 
 
 def fft(a, n=None, axis=-1, norm=None):
@@ -24,7 +158,7 @@ def fft(a, n=None, axis=-1, norm=None):
 
     .. seealso:: :func:`numpy.fft.fft`
     """
-    return cufft.fft(a, (n,), (axis,), norm, cupy.cuda.cufft.CUFFT_FORWARD)
+    return _fft(a, (n,), (axis,), norm, cupy.cuda.cufft.CUFFT_FORWARD)
 
 
 def ifft(a, n=None, axis=-1, norm=None):
@@ -45,7 +179,7 @@ def ifft(a, n=None, axis=-1, norm=None):
 
     .. seealso:: :func:`numpy.fft.ifft`
     """
-    return cufft.fft(a, (n,), (axis,), norm, cupy.cuda.cufft.CUFFT_INVERSE)
+    return _fft(a, (n,), (axis,), norm, cufft.CUFFT_INVERSE)
 
 
 def fft2(a, s=None, axes=(-2, -1), norm=None):
@@ -66,7 +200,7 @@ def fft2(a, s=None, axes=(-2, -1), norm=None):
 
     .. seealso:: :func:`numpy.fft.fft2`
     """
-    return cufft.fft(a, s, axes, norm, cupy.cuda.cufft.CUFFT_FORWARD)
+    return _fft(a, s, axes, norm, cufft.CUFFT_FORWARD)
 
 
 def ifft2(a, s=None, axes=(-2, -1), norm=None):
@@ -87,7 +221,7 @@ def ifft2(a, s=None, axes=(-2, -1), norm=None):
 
     .. seealso:: :func:`numpy.fft.ifft2`
     """
-    return cufft.fft(a, s, axes, norm, cupy.cuda.cufft.CUFFT_INVERSE)
+    return _fft(a, s, axes, norm, cufft.CUFFT_INVERSE)
 
 
 def fftn(a, s=None, axes=None, norm=None):
@@ -108,7 +242,7 @@ def fftn(a, s=None, axes=None, norm=None):
 
     .. seealso:: :func:`numpy.fft.fftn`
     """
-    return cufft.fft(a, s, axes, norm, cupy.cuda.cufft.CUFFT_FORWARD)
+    return _fft(a, s, axes, norm, cufft.CUFFT_FORWARD)
 
 
 def ifftn(a, s=None, axes=None, norm=None):
@@ -129,7 +263,7 @@ def ifftn(a, s=None, axes=None, norm=None):
 
     .. seealso:: :func:`numpy.fft.ifftn`
     """
-    return cufft.fft(a, s, axes, norm, cupy.cuda.cufft.CUFFT_INVERSE)
+    return _fft(a, s, axes, norm, cufft.CUFFT_INVERSE)
 
 
 def rfft(a, n=None, axis=-1, norm=None):
@@ -151,8 +285,7 @@ def rfft(a, n=None, axis=-1, norm=None):
 
     .. seealso:: :func:`numpy.fft.rfft`
     """
-    return cufft.fft(a, (n,), (axis,), norm, cupy.cuda.cufft.CUFFT_FORWARD,
-                     'R2C')
+    return _fft(a, (n,), (axis,), norm, cufft.CUFFT_FORWARD, 'R2C')
 
 
 def irfft(a, n=None, axis=-1, norm=None):
@@ -176,8 +309,7 @@ def irfft(a, n=None, axis=-1, norm=None):
 
     .. seealso:: :func:`numpy.fft.irfft`
     """
-    return cufft.fft(a, (n,), (axis,), norm, cupy.cuda.cufft.CUFFT_INVERSE,
-                     'C2R')
+    return _fft(a, (n,), (axis,), norm, cufft.CUFFT_INVERSE, 'C2R')
 
 
 def rfft2(a, s=None, axes=(-2, -1), norm=None):
@@ -199,7 +331,7 @@ def rfft2(a, s=None, axes=(-2, -1), norm=None):
 
     .. seealso:: :func:`numpy.fft.rfft2`
     """
-    return cufft.fft(a, s, axes, norm, cupy.cuda.cufft.CUFFT_FORWARD, 'R2C')
+    return _fft(a, s, axes, norm, cufft.CUFFT_FORWARD, 'R2C')
 
 
 def irfft2(a, s=None, axes=(-2, -1), norm=None):
@@ -223,7 +355,7 @@ def irfft2(a, s=None, axes=(-2, -1), norm=None):
 
     .. seealso:: :func:`numpy.fft.irfft2`
     """
-    return cufft.fft(a, s, axes, norm, cupy.cuda.cufft.CUFFT_INVERSE, 'C2R')
+    return _fft(a, s, axes, norm, cufft.CUFFT_INVERSE, 'C2R')
 
 
 def rfftn(a, s=None, axes=None, norm=None):
@@ -245,7 +377,7 @@ def rfftn(a, s=None, axes=None, norm=None):
 
     .. seealso:: :func:`numpy.fft.rfftn`
     """
-    return cufft.fft(a, s, axes, norm, cupy.cuda.cufft.CUFFT_FORWARD, 'R2C')
+    return _fft(a, s, axes, norm, cufft.CUFFT_FORWARD, 'R2C')
 
 
 def irfftn(a, s=None, axes=None, norm=None):
@@ -269,7 +401,7 @@ def irfftn(a, s=None, axes=None, norm=None):
 
     .. seealso:: :func:`numpy.fft.irfftn`
     """
-    return cufft.fft(a, s, axes, norm, cupy.cuda.cufft.CUFFT_INVERSE, 'C2R')
+    return _fft(a, s, axes, norm, cufft.CUFFT_INVERSE, 'C2R')
 
 
 def hfft(a, n=None, axis=-1, norm=None):
